Crate latexml_core

The Core of latexml - roughly the equivalent of TeX conversion.

Re-exports

pub use crate::common::error::*;

Modules

alignment

Support for TeX-like Alignments

aux_macros

Auxiliary macros Helper macros for quicker and more idiomatic construction and access to data structures.

binding

The programmable API foundation for creating bindings of LaTeX sty/cls libraries

comment

TeX comments as standalone objects

common

Common abstractions that are useful at various stages of Core processing

definition

All possible definitions for TeX-native commands (expandable, primitive, constructor,…)

digested

a shared interface for digested objects Interface layer for the full range of digested objects

document

An abstraction layer over the converted XML document

gullet

The Gullet is responsible for reading Tokens and other data from the Mouth

keyval

A LaTeX-like Key-Value object Key-Value Definitions

keyvals

A collection of Key-Value objects, typically from a single LaTeX argument

ligature

Rules for combining together characters and other text rules

list

A list of Digested objects

mouth

The mouth is a thin interface over a file, responsible for reading characters and associating them with catcodes

parameter

The abstraction layer used by the Gullet to read arguments for the various kinds of TeX object definitions

rewrite

Rules for rewriting the constructed XML document, after core processing has completed

state

A global, singleton, mutable state - hosts almost all TeX-facing runtime information for the conversion

stomach

The stomach is an abstraction responsible for digesting Tokens and Registers prepared by the Gullet into Boxes

tbox

A TeX-like digested Box

token

The Token and its Catcode

tokens

Grouping Tokens together Token List constructors.

util

Auxilary utilities that do not participate in the main conversion abstraction

whatsit

A TeX-like digested Whatsit

Macros

AssignRegister

CharToken

Special case: a character needs swift string conversion, so let’s use a dedicated macro

Debug

Error

Explode

Explode a string into a list of tokens, all w/catcode OTHER (except space).

ExplodeChars

ExplodeText

Similar to Explode, but convert letters to catcode LETTER and others to OTHER Hopefully, this is essentially correct WITHOUT resorting to catcode lookup?

Fatal

Font

build a Font from key=>val pairs

FontDirective

build a FontDirective from key=>val pairs (currently only FontDirective::Asset is supported in this macro)

Info

NewDefault

given a struct $name, create a new instance of it using the given key=>val pairs and complete the remaining entries via the Default instance

NewDefaultV

Just like NewDefault, but adds a mandatory .into_option() to all values

Number

Roman

Convert the number to upper case roman numerals, returning a list of LaTeXML::Core::Token

SetCounter

SymExplodeText

T_ACTIVE

macro for an ACTIVE char token

T_ALIGN

macro for an ALIGN “&” token

T_ARG

macro for a numbered ARG token

T_BEGIN

macro for a BEGIN “{” token

T_COMMENT

macro for a COMMENT content token

T_CR

macro for a CR “\n” token

T_CS

macro for a command sequence token

T_END

macro for a new END “{” token

T_LETTER

macro for a LETTER token

T_MARKER

macro for a tracing MARKER token

T_MATH

macro for a MATH “$” token

T_OTHER

macro for an OTHER code token

T_OTHER_CHAR

T_OTHER from a single character

T_PARAM

macro for a PARAM “#” token

T_RELAX

macro for T_CS(“\relax”)

T_SPACE

macro for a SPACE token (default “ “)

T_SUB

macro for a SUB “_” token

T_SUPER

macro for a SUPER “^” token

Token

Token constructor macro (defaults to OTHER code)

Tokens

convenience macro for assembling a Tokens object from different pieces (Token, Vec<Token>, Tokens)

Warn

after_digest

after_digest_simple

after_digest_single

before_digest

before_digest_simple

before_digest_single

construct

count_unpack_ref

fatal

fontmap

Instantiates a Font, using the Font fields as keys, and calling .into() for each value. The specification can be partial - missing fields are taken via the Default trait.

forbidMath

generate_message

getter

A closure for obtaining a RegisterValue, usually owned by a Register getter.

map

A flexary macro for constructing HashMap<String, T> maps, where T is generic.

mixrc

A variant on mixvec! where each argument receives an additional .into() call best used with an outer context that explicitly provides the expected type, such as

mixvec

A variant on vec! where each argument receives an additional .into() call best used with an outer context that explicitly provides the expected type, such as

noreplacement

predigest

print_stderr

prints a to STDERR without a line break

println_stderr

prints a single line to STDERR

prop_bool

prop_digested

prop_str

prop_string

prop_whatsit

properties

raw_char_map

A flexary macro for constructing HashMap<char, T> maps, where T is generic

raw_map

A flexary macro for constructing HashMap<K, V> maps, where K and V are both generic (inferred at time of use)

reader

replacement

report

report_mut

requireMath

reversion

reversion_digested

rewrite_replace_sub

roman

Convert the number to lower case roman numerals, returning a list of LaTeXML::Core::Token

s

The s! macro is a briefer alias for format!

set

Simple generic helper for HashSet<K,V> creation Source: https://riptutorial.com/rust/example/4149/create-a-hashset-macro

setter

sizersub

some

The some! macro transforms data in type S to Option<Into<T>> (always wrapping with Some)

static_map

A flexary macro for constructing HashMap<&'static str, &'static str> maps

stored_map

A flexary macro for constructing SymHashMap<Stored> maps

string_keys_map

A flexary macro for constructing HashMap<String, T> maps

string_map

A flexary macro for constructing HashMap<String, String> maps

sym_map

A flexary macro for constructing SymHashMap<T> maps

tagsub

transfer_opt_default

Transfers a mutable pointer to a hashmap entry, or fills in with a default if missing.

unpack_opt

Try to efficiently unwrap a Vec<T> into a [T;n] for $arg1…$argn

unpack_opt_ref

Try to efficiently unwrap a &Vec<T> into a &[T;n] for $arg1…$argn

unported

unref

Convenience macro to flexibly unpack a collection of Vec<ArgWrap> arguments into individual Tokens variables.

Structs

Core

The Core conversion runtime

CoreOptions

Configuration for the Core processing

Enums

TexMode

The current TeX processing mode

Statics

NO_PROPERTIES

Traits

BoxOps

Common operations for Box-like (digested) data