latexml_core::binding::def::dialect

Function def_constructor

pub fn def_constructor(
    cs: Token,
    paramlist: Option<Parameters>,
    compiled_replacement: Option<ReplacementClosure>,
    options: ConstructorOptions,
)

Binding definition connecting a TeX command sequence with a structured XML output.

The Constructor is where LaTeXML really starts getting interesting; invoking the control sequence will generate an arbitrary XML fragment in the document tree. More specifically: during digestion, the arguments will be read and digested, creating a Whatsit to represent the object. During absorption by the Document, the Whatsit will generate the XML fragment according to the compiled_replacement.