latexml_core::binding::def::dialect

Function def_math

pub fn def_math(
    cs: Token,
    paramlist: Option<Parameters>,
    presentation: String,
    options: MathPrimitiveOptions,
) -> Result<()>

Define a Mathematical symbol or function.

There are two sets of cases: (1) If the presentation appears to be TeX code, we create an XMDual, since the presentation may end up with structure, etc. (2) But if the presentation is a simple string, or unicode, it is just the content of the symbol; even if the function takes arguments.