latexml_core::binding::def::dialect

Function def_primitive

pub fn def_primitive(
    cs: Token,
    paramlist: Option<Parameters>,
    compiled_replacement: Option<PrimitiveBody>,
    options: PrimitiveOptions,
) -> Result<()>

Defines a primitive control sequence

A primitive is processed during digestion (in the Stomach), after macro expansion but before Construction time. Primitive control sequences generate Boxes or Lists, generally containing basic Unicode content, rather than structured XML. Primitive control sequences are also executed for side effect during digestion, effecting changes to the State.