latexml_core::common
Module local_assignments

Perl-style local assignments (dynamic scope)

This module can benefit from some thinking over and refactoring in principle the “local” scope variables in Perl are a completely standalone feature (it’s global namespace value shadowing until scope expiration) … so the push/pop can be modeled with sentry values getting created and dropped (TODO)…

Structs

Localized

These are fields realized via Perl’s “local” mechanism in LaTeXML, but (for now) require explicit “expire” calls in Rust. Ideally their ergonomics gets improved, or we gradually phase them out.

Functions

align_group_count

decrement_align_group_count

expire_align_group_count

expire_current_token

expires the most recent (localized) current token.

expire_dual_branch

expire (localized) flag for “dual branch”

expire_ifframe

expires the most recent (originally Perl-local) IfFrame

expire_reading_alignment

expire_state_unlocked

get_current_token

gets the (localized) current token

get_dual_branch

get the current value for “dual branch”

get_ifframe

retrieves the most recent (originally Perl-local) IfFrame

get_reading_alignment

has_reading_alignment

increment_align_group_count

local_align_group_count

local_build_template

local_current_token

localizes a new current token. see Stomach::invoke_token

local_reading_alignment

local_state_unlocked

set_align_group_count

set_build_template

set_dual_branch

sets the (localized) flag for “dual branch”

set_ifframe

sets a (originally Perl-local) IfFrame that needs to be manually expired.

state_is_unlocked

take_build_template

with_current_build_template