latexml_core
Module gullet

The Gullet is responsible for reading Tokens and other data from the Mouth

Structs

Gullet

MouthRuntime

Enums

ExpansionLevel

A directive describing to what degree a gullet reader should perform TeX’s expansion

Statics

GULLET

Functions

close_mouth

do_expand

do_expand_partially

flush

Obscure, but the only way I can think of to End!! (see \bye or \end{document}) Flush all sources (close all pending mouth’s)

flush_mouth

This flushes a mouth so that it will be automatically closed, next time it’s read Corresponds to TeX’s \endinput

get_location

Comment-oriented location string, based on get_locator

get_locator

Get the current location of input getting read

get_pending_comment

has_more_input

Check if there is more input to be read from the current mouth

if_next

initialize_gullet

Initialize (or reset, if reentrant) a Gullet to its default empty state

is_column_end

mouth_is_open

open_mouth

read_arg

Read and return a “normal” TeX argument

read_balanced

Approximates TeX’s scan_toks (but doesn’t parse \def parameter lists) and only optionally requires the openning “{”.

read_cs_name

read_dimension

read_float

====================================================================== Float, a floating point number. Similar to factor, but does NOT accept comma! This is NOT part of TeX, but is convenient.

read_glue

read_keyword

Match the input against a set of keywords; Similar to readMatch, but the keywords are strings, and Case and catcodes are ignored; additionally, leading spaces are skipped. AND, macros are expanded.

read_match

Match the input against one of the Token or Tokens in @choices; return the matching one or undef.

read_mu_dimension

read_mu_glue

read_mu_unit

read_next_conditional

reads and discards tokens, until it encounters a conditional, if any

read_non_space

Read a single non-space token

read_normal_integer

read_number

read_optional

Read and return a LaTeX optional argument

read_raw_line

Read the next raw line (string); primarily to read from the Mouth, but keep any unread input!

read_register_value

read_rubber

read_token

read_tokens_value

Apparent behaviour of a token value (ie \toks#=<arg>)

read_until

Return a (balanced) sequence tokens until a match against one of the Tokens in @delims.

read_until_brace

reads until it encounters a Catcode::BEGIN token

read_until_token

Convenience method wrapping around read_until TODO: This seems to be the wrong Rust type interface, we need to rework…

read_value

read_x_non_space

Read a single expanded, non-space, token

read_x_token

Read the next non-expandable token (expanding tokens until there’s a non-expandable one).

reading_from_mouth

Do something, while reading tokens from a specific Mouth.

skip_filler

= | \relax TeX Book p.276 “ can be implicit”, and experimentation, indicate Expansion!!!

skip_one_space

skip_spaces

unread

Push the tokens back into the input stream to be re-read.

unread_mut

Variant of unread, but drains the contents of tokens without taking ownership.

unread_one

Unreads a single Token to the start of the token stream

unread_vec

Unreads a Vec<Token> to the start of the token stream

with_mouth_mut

Execute a function with a mutable reference to the current mouth