latexml_core
Module mouth

The mouth is a thin interface over a file, responsible for reading characters and associating them with catcodes

Structs

Mouth

MouthOptions

Enums

FoodType

Functions

tokenize

tokenize_internal