latexml_core

Module state

A global, singleton, mutable state - hosts almost all TeX-facing runtime information for the conversion

Re-exports

pub use crate::common::store::Stored;

pub use crate::common::local_assignments::*;

Structs

State

The state efficiently bookkeeps the bindings in a TeX-like fashion.

StateOptions

state fields allowed for customization during construction

UndoFrame

For each of several tables (being “value”, “meaning”, “catcode” or other space of names), each table maintains the bound values, and “undo” defines the stack frames

Enums

Catcodes

High-level catcode profiles

Scope

installation scope in the state_tables

TableName

the kinds of tables bookkept in the State

Statics

UNITS

Used in conversion to scaled points

Functions

activate_scope

Activates all stashed definitions for the named scope. No-op if the scope is already active.

add_search_path

after_assignment

run the accumulated directives from \afterassignment

assign_alignment

assign_catcode

assigns a Catcode for a given character

assign_delcode

like assign_catcode but targets Delcode and its table

assign_font

a convenience method to globally asign a Font to the “font” key

assign_lccode

like assign_catcode but targets Lccode and its table

assign_mapping

assign_mathcode

like assign_catcode but targets Mathcode and its table

assign_meaning

$meaning should be a definition (for defining active control sequences) or another token, for \let

assign_register

assign_sfcode

like assign_catcode but targets Sfcode and its table

assign_uccode

like assign_catcode but targets Uccode and its table

assign_value

assigns a Stored value at the given key and scope

assign_value_sym

assigns a Stored value at the given (arena ticket!) key and scope

begin_semiverbatim

begins a semiverbatim frame, neutralizing the usual + requested characters

checkin_value

Returns a value into its Stored::None placeholder (see checkout_value for taking it)

checkout_value

Replaces the value in question with Stored::None (see checkin_value for returning it)

clear_prefixes

clears the global prefixes

compute_indirect_model

The indirect model includes all elements allowed as direct children, and all descendents of a node that can be inserted after autoOpen’ing intermediate elements.

convert_unit

convert a unit name into a f64 scaling factor over sp

current_verbosity

deactivate_scope

Removes any definitions that were associated with the named scope. Normally not needed, since a scopes definitions are locally bound anyway.

end_semiverbatim

end by just calling pop_frame

ensure_tag_property

generate_error_stub

Generate a stub definition for an undefined control-sequence, along with appropriate error messge.

generate_ligature_id

simple id generator for a ligature

get_active_scopes

return the currently activated named scopes

get_bindings_dispatch

get_extra_bindings_dispatch

get_frame_depth

Determine depth of group nesting.

get_graphics_paths

get_indirect_model_relationship

get_input_encoding

get_known_scopes

return all known named scopes

get_nomathparse_flag

get_prefix

gets the current value of a named prefix

get_search_paths

get_tag_property

graphics_paths_push_front

has_indirect_model

has_meaning

has_search_paths

has_value

Check if the Value table contains a given key

install_definition

A shorthand for installing definitions

is_dont_expandable

Whether token is affected by \noexpand

is_value_bound

Was name bound? If frame is given, check only whether it is bound in that frame (0 is the topmost).

let_i

Let macro setter

lookup_alignment

lookup_bool

A bit of Perl “existence as truth” semantics mixed in with proper boolean lookup

lookup_catcode

Lookup & assign a character’s Catcode

lookup_conditional

lookup_definition

used for expansion & various queries Since we’re not doing digestion here, we don’t need to handle mathactive, nor cs let to executable tokens This returns a definition object, or undef

lookup_definition_stored

Returns a definition as Stored so that one can call .read_arguments

lookup_delcode

like lookup_catcode but targets Delcode and its table

lookup_digestable_definition

Recognizes mathactive tokens in math mode and also looks for cs that have been let to other `executable’ tokens. Returns a definition object, or a “self inserting” token. Used for digestion.

lookup_dimension

a variant of lookup_value that casts the value into Dimension

lookup_expandable

lookup_font

convenience method to lookup the current value at the “font” key

lookup_glue

a variant of lookup_value that only recognizes a Stored::Glue

lookup_int

like lookup_value but only recognizes Int, Bool and Number variants of Stored (default: 0)

lookup_lccode

like lookup_catcode but targets Lccode and its table

lookup_mapping

lookup_mathcode

like lookup_catcode but targets Mathcode and its table

lookup_mathcode_sym

lookup_mathfont

convenience method to lookup the current value at the “mathfont” key

lookup_meaning

Get the `Meaning’ of a token. For active control sequences this may give the definition object (if defined) or another token (if \let) or undef Any other token is returned as is.

lookup_muglue

a variant of lookup_value that only recognizes a Stored::Glue

lookup_number

a variant of lookup_value that casts the value into Number

lookup_register

lookup_register_definition

A specialized version of lookup_definition for registers, since we can’t adequately perform multi-dispatch when we have a “Self: Sized” for the Definition trait object.

lookup_sfcode

like lookup_catcode but targets Sfcode and its table

lookup_string

like lookup_value, but casts the entry into a String (empty if None)

lookup_string_sym

like lookup_value, but casts the entry into a SymStr from the string interner (EMPTY_SYM if None)

lookup_token

a variant of lookup_value that only recognizes a Stored::Token

lookup_tokens

a variant of lookup_value that casts the response into Tokens

lookup_uccode

like lookup_catcode but targets Uccode and its table

lookup_value

lookup_vecdeque

like lookup_value but only recognizes Stored::VecDequeStored

pop_frame

Ends the current level of grouping. Note that this is lower level than \egroup;

pop_value

pops the last value in a named Stored::VecDequeStored queue, if any

push_frame

Starts a new level of grouping. Note that this is lower level than C<\bgroup>;

push_pending_resource

push_tokens

Pushes Tokens into a Stored::Tokens value when defined, or assigns when new.

push_value

manage a (global) list of values

remove_value

inline lookup_value after which globally assign an empty Tokens() to undo

remove_vecdeque

reset_pending_resources

search_paths_push_front

set_bindings_dispatch

set_extra_bindings_dispatch

set_indirect_model

set_input_encoding

set_nomathparse_flag

set_prefix

Set one of the definition prefixes global, etc (only global matters!)

set_state

shift_value

take_pending_resources

unshift_value

use_main_state

use_std_state

use_sty_state

with_font_info

with_mapping

manage a (global) hash of values

with_mapping_keys

with_mapping_sym

with_search_paths

with_stacked_values

with_tag_property

with_tag_property_mut

with_value

with_value_mut

with_vecdeque

x_equals

XEquals check for two token arguments

Type Aliases

AssignmentCount

Ledger for stacked assignments

StashTable

The (table_name, key, value) contents of a stored table of assignments

Table

The type of values that are storable by the different namespaced “tables” in State.