latexml_core::tbox

Struct Tbox

pub struct Tbox {
    pub text: SymStr,
    pub font: Rc<Font>,
    pub locator: Option<Locator>,
    pub properties: SymHashMap<Stored>,
    pub tokens: Tokens,
}

Box is a Rust keyword, so we use “Tbox” instead, as in “TeX Box”

Fields
text: SymStr

plain-text content

font: Rc<Font>

associated font for text

locator: Option<Locator>

source location where the box originated (None = unknown)

properties: SymHashMap<Stored>

misc properties, such as sizing information

tokens: Tokens

a Tokens list containing the TeX that created (or could have) the Tbox.

Implementations

impl Tbox

pub fn new( text: SymStr, font_opt: Option<Rc<Font>>, locator_opt: Option<Locator>, tokens_opt: Tokens, properties: HashMap<Stored>, ) -> Self

creates a new Tbox. If font_opt or locator_opt are None, they are obtained from the currently active state? Note that text can be empty, which contributes nothing to the generated document, but does record the TeX code (in the tokens).

pub fn is_empty(&self) -> bool

checks if the text content is empty

pub fn is_math(&self) -> bool

Whether this box is in math mode. Perl: Box.pm::isMath (L79-81): ($mode || 'restricted_horizontal') =~ /math$/. Matches “math”, “inline_math”, “display_math”.

pub fn set_properties<I>(&mut self, entries: I)
where I: IntoIterator<Item = (&'static str, Stored)>,

Batch-insert properties. Equivalent to calling set_property for each entry, but avoids re-looking up the properties map per call. Perl: Box.pm::setProperties (L171-175).

pub fn total_height(&self) -> Dimension

height + depth of this box (the TeX “total height”). Returns Dimension::default() (0pt) when either component is absent. Perl: Box.pm::getTotalHeight (L202-210).