latexml_core::token

Struct Token

pub struct Token {
    pub text: SymStr,
    pub code: Catcode,
}

The core immutable syntactic primitive resulting from TeX’s read-in and expansion process We allow the fields to be public, so that we can use builder macros such as

macro_rules! T_SPACE(() => {
    Token { text: arena::pin_static(" "), code: Catcode::SPACE}
  });

Fields
text: SymStr

an arena id the character content for this token

code: Catcode

a TeX catcode

Implementations

impl Token

====================================================================== Accessors.

pub fn new<T: AsRef<str>>(text: T, code: Catcode) -> Self

simple Token constructor, wrapping over text and catcode

pub fn get_cs_name(&self) -> SymStr

Get the CS Name of the token. This is the name that definitions will be stored under; It’s the same for various `different’ BEGIN tokens, eg.

pub fn with_cs_name<R, FnR>(&self, caller: FnR) -> R
where FnR: FnOnce(&str) -> R,

Execute a closure using the CS Name of the token. This is the name that definitions will be stored under; It’s the same for various `different’ BEGIN tokens, eg.

pub fn pin_cs_name(&self) -> SymStr

artificial, but avoids the data race

pub fn get_primitive_name(&self) -> Option<&'static str>

Get the fixed name of a primitive catcode, or empty string otherwise

pub fn get_executable_name(&self) -> String

Get the CS name only if the catcode is executable!

pub fn get_executable_primitive_name(&self) -> Option<&'static str>

Intersect executable and primitive

pub fn get_sym(&self) -> SymStr

Use the ticket representing the interned “text” of the token

pub fn with_str<R, FnR>(&self, caller: FnR) -> R
where FnR: FnOnce(&str) -> R,

Use the interned &str “text” of the token use to_string instead for an owned String with simpler

pub fn get_charcode(&self) -> u32

Return the character code of character part of the token, or 256 if it is a control sequence

pub fn get_catcode(&self) -> Catcode

Return the catcode of the token.

pub fn is_executable(&self) -> bool

is the current one

pub fn neutralize(self, extraspecials: &[char]) -> Token

neutralize really should only retroactively imitate what Semiverbatim would have done. So, it needs to neutralize those in SPECIALS NOTE that although ‘%’ gets it’s catcode changed in Semiverbatim, I’m pretty sure we do NOT want to neutralize comments (turn them into Catcode::OTHER) here, since if comments do get into the Tokens, that will introduce weird crap into the stream.

pub fn as_other(&self) -> Token

pub fn as_cs(&self) -> Token

pub fn substitute_parameters(self, args: &[&Token]) -> Self

pub fn revert(self) -> Token

A Token reverts to itself

pub fn stringify(&self) -> String

A string form which is primarily used for error-reporting

pub fn to_register(&self) -> Option<Rc<Register>>

pub fn to_number(&self) -> Number

pub fn to_dimension(&self) -> Dimension

pub fn to_mu_dimension(&self) -> MuDimension

pub fn to_glue(&self) -> Glue

pub fn to_mu_glue(&self) -> MuGlue

pub fn to_float(&self) -> Float

pub fn be_digested(self) -> Result<Digested>

pub fn defined_as(&self, other: &Token) -> bool

Check whether the current token is defined as other. That is, whether it is equal to other, or \let to other. other is is presumed to be some “constant”, explicit token, such as T_SPACE or T_CS!("\\endcsname").