latexml_core::common
Module font

Structs

Font

A data structure containing Font information, but also related textual properties (such as color)

Constants

MATH_FONTS

TEXT_FONTS

Statics

FONT_TEXT_DEFAULT

Functions

decode

Decode a codepoint using the fontmap for a given font and/or fontencoding.

decode_fontname

decode_string

lookup_font_family

Global auxiliary for font family lookup

lookup_font_series

Global auxiliary for font series lookup

lookup_font_shape

Global auxiliary for font shape lookup

rationalize_font_size

Convert stanard font size names, such as tiny, Huge, etc to f64

relative_font_size

convert size to percent

Type Aliases

Fontmap