latexml_core::definition

Module register

Structs

CharDefProps

Register

A struct representing a TeX register

Enums

RegisterType

The type of a TeX Register

RegisterValue

The values that can be read by, and stored in, a Register

Type Aliases

RegisterGetterClosure

looks up a stored value from the state frame (at a constant key, or key based on the arguments)

RegisterSetterClosure

sets a register value in the state::frame